+++ node/directives/validateStoreUserAccess.ts
@@ -41,13 +41,16 @@ export class ValidateStoreUserAccess extends SchemaDirectiveVisitor {
         userAgent,
       }
 
-      const { hasAdminToken, hasValidAdminToken } = await validateAdminToken(
-        context,
-        adminUserAuthToken as string
-      )
+      const { hasAdminToken, hasValidAdminToken, hasValidAdminTokenFromStore } =
+        await validateAdminToken(context, adminUserAuthToken as string)
 
       // add admin token metrics
-      metricFields = { ...metricFields, hasAdminToken, hasValidAdminToken }
+      metricFields = {
+        ...metricFields,
+        hasAdminToken,
+        hasValidAdminToken,
+        hasValidAdminTokenFromStore,
+      }
 
       // allow access if has valid admin token
       if (hasValidAdminToken) {
@@ -63,13 +66,15 @@ export class ValidateStoreUserAccess extends SchemaDirectiveVisitor {
         return resolve(root, args, context, info)
       }
 
-      const { hasApiToken, hasValidApiToken } = await validateApiToken(context)
+      const { hasApiToken, hasValidApiToken, hasValidApiTokenFromStore } =
+        await validateApiToken(context)
 
       // add API token metrics
       metricFields = {
         ...metricFields,
         hasApiToken,
         hasValidApiToken,
+        hasValidApiTokenFromStore,
       }
 
       // allow access if has valid API token